Project set up
Imported and modified one of my scripts that I was working on for this game (All I did was make a small start)
With this movement script you can move left and right, jump, walljump dash, sprint and I added a ledgegrab mechanic.
I also made the script so it uses states.

I created an input action map for the player movement , added a wall jump tag and made a player prefab.