New state and some blocking logic
Added a combat state so it is easier to differentiate between what happening in movement and the action the player is taking.

i tried to fix the ledgegrabbing so you can ledgegrab even if the ledge is behind you, but I didn't know a good solution and jsut hard coded it, but I didnt finish the last step.

I added a blocking function, you can now go into a blocking state, but I struggle with returning to a neutral position since ctx.cancelled doesn't get called for some reason.