More optimized player controller and start to local multiplayer managers
Organized the files a little more.
Added more consistent logic in player controller, still need to add more logic to the launch and hit detection since those are a little iffy now.

changing the structure of the player and local multiplayer. A player class get instantiated with a player ID, it parses all the inputs into the playerController and the player class itself will have the property to navigate UI. there is a player manager that keeps track of everything and should handle more of the backend.

also with the new unity auditor profiler thing, we tested something out and fixxed all bitmap texture "issue's", that is mostly causing all of these changed files