Start to animation setup
Made and set up the animator for the character, some parameters still might need to be assigned. Also need to set the parameters through the character controller script.

made a new prefab of the character model with all the necessary components atached
fixxed an issue where the player would always be in the in air state.
player cant hit itself anymore.