Beginning for the drawing mechanic
Made a beginning to the drawing mechanic in the painter character class. You should theoretically be able to draw some points in a 3D environment and with an ear clipping algorithm (that chatgpt wrote for me) the whole shape triangulates and you can use that data to generate a mesh.

It still needs alot of fail saves an play testing when it is finished. Also need to make it so you can give the collider different property's