Changed how mesh was created
Instead of having only 1 collider for the whole drawing, the drawing is divided in multiple colliders. Because a concave mesh cant be used as a trigger I had to find a different way to give it the same feel as if the drawing itself is the trigger. It can probably get optimized but I dont have the time for that now.