Heavy attacks animations and draw ability test
Better implemented the charge/heavy attacks. Still have to fix the animations for them.

Since I wasn't satisfied with how the triangulation and collision worked with the mesh I tried to test some different algolrithms. I now found an algorithm that works fine.
it is in the ConvexDecomposition class. it still has it's issues that I need to look into later.
It creates a bunch of smaller convex meshes but it triangulates the opposite way so the drawing is turned away from the camera. it also most of the time creates one big hull convex collider that defeats the purpose of all the smaller ones