Drawing optimisation
Went back to an older triangulation method ear clipping. instead of making several compound colliders or different meshes with mesh colliders.

You now only create one mesh with a 2d polygoncollider and instead of looking for collisions with colliders. I use collider.ClosestPoint to check the distance between the players and the drawn mesh.

The mesh now gets a proper pivot point, wich is used for making a second mesh that is a little bit bigger to be used as a background. I still need to change its transform a little but because it generated 2 meshes on top of eachother now.

Added small amounts of logic for choosing which property the player wants to pick out of 4 property's