Attack number tweaks and changed knockback
Some attack animations are a little faster now, there are no transitions times anymore. Changed some timers in attackdata. But it still feels really janky.

Changed how knockback gets applied and it feels alot better. I still need to change gravity logic a little. I need it to scale a little bit with the knockback intensity.

Added the Property select wheel to the player and the functionality for it. Only need to enable it when ulting and not constantly enabled.