Changed how attacks get detected
Attacks instead of overlapbox, just simply get detected by an ontrigger function that is on the brush collider. (it still not finished, I need to clear up some things and do the same for the heavy attacks)

Also changed some player attack values (attackStart time and some gravity tweaks). Made an attackcolliders script that is attached to the brush/collider that needs the hit detection